PLEASE TRY SETTING:
PlanetMap.ISOMETRIC = true  [line 19](if it is false it will just be as
normal)

Only stuff I have been playing around with and working out the dynamics.
far from look perfect but a start or something that has showed a lot of
the problems we will face when we implement it.

Also, i havent been able to figure out how to transfer the squad
coordinates into the Isometric tile coordinates... so the squad is not
responding to the underlying terrain or speed at the moment.